cleanup: remove double 'use client', move useCanvasLoop to hooks/canvas (#319)
- Remove duplicate 'use client' from useSpeechSynthesis, useGameEvents,
  useAutoGame, and useGameContext (kept only the top-of-file occurrence)
- Move useCanvasLoop from hooks/shared/common/ to hooks/canvas/ alongside
  useCanvasReady; add hooks/canvas/index.ts barrel
- Update all 10 game hook callers to import from @/hooks/canvas
- Delete now-empty hooks/shared/common/ directory

Closes #319
