+++ gamesformykids/app/games/pong/usePongGame.ts
@@ -3,7 +3,7 @@
 import { useEffect, useRef, useCallback } from 'react';
 import { useShallow } from 'zustand/react/shallow';
 import { usePongStore } from './pongStore';
-import { useCanvasLoop } from '@/hooks/shared/common';
+import { useCanvasLoop } from '@/hooks/canvas';
 
 export const W = 360;
 export const H = 560;
