+++ gamesformykids/hooks/canvas/index.ts
@@ -1,0 +1,2 @@
+export { useCanvasLoop } from './useCanvasLoop';
+export { useCanvasReady } from './useCanvasReady';
