+++ gamesformykids/hooks/shared/audio/useSpeechSynthesis.ts
@@ -1,7 +1,5 @@
 'use client';
 
-'use client';
-
 import { useCallback } from 'react';
 import { AUDIO_CONSTANTS } from '../../../lib/constants/core';
