--- gamesformykids/hooks/shared/common/index.ts
@@ -1,7 +1,0 @@
-/**
- * ===============================================
- * Shared Common Hooks - Index
- * ===============================================
- */
-
-export { useCanvasLoop } from './useCanvasLoop';
