+++ gamesformykids/hooks/shared/ui/useGameEvents.ts
@@ -1,10 +1,3 @@
-/**
- * Game Events Hook
- * Hook לניהול אירועי משחק ותגובות
- */
-
-'use client';
-
 'use client';
 
 import { useCallback, useEffect, useMemo } from 'react';
